Solve day 16
Tried so many approaches and iterations before this .. The final
solution takes some inspiration from this message on Reddit:

> My understanding is that basically what it's doing is division of
> labor. Basically, I'm responsible for 15-X valves, the elephant is
> responsible for the other X, we both try to get our parts done as fast
> as possible.
>
> Then looking at all 2^14 different ways to split up 15 valves between
> two individuals gets you an answer.

I had all the dynamic programming plumbing (and a lot more) in place,
but I was going through it differently. The program even calculated
answers in reasonable time at different points but was giving the wrong
answer to the real input for part 2. This one now worked for all parts
and inputs.