Solve day 22 2024
I actually had most of it solved during the day of, but something wasn't
working for part two. Turns out it was the fact that I was using
`HM.fromListWith const`, which meant that the value of the last matching
sequence was used, not the first one as required by the problem.
Swapping it out for the equivalent of `flip const` was all that was
needed to fix it.